[JAX-WS-807] * Making T_END not zero to capture any any buffer overruns * Exposing a lowlevel method to write an structure element for T_END Issue number: Obtained from: Submitted by: Reviewed by:
svn path=/trunk/; revision=302

+++ streambuffer/src/com/sun/xml/stream/buffer/MutableXMLStreamBuffer.java
@@ -91,7 +91,7 @@ public class MutableXMLStreamBuffer extends XMLStreamBuffer {
 
         // Set the first element of structure array to indicate an empty buffer
         // that has not been created
-        _structure.getArray()[0] = AbstractCreatorProcessor.T_END;
+        _structure.getArray()[0] = (byte) AbstractCreatorProcessor.T_END;
     }
 
     /**
@@ -213,7 +213,7 @@ public class MutableXMLStreamBuffer extends XMLStreamBuffer {
 
         // Set the first element of structure array to indicate an empty buffer
         // that has not been created
-        _structure.getArray()[0] = AbstractCreatorProcessor.T_END;
+        _structure.getArray()[0] = (byte)AbstractCreatorProcessor.T_END;
 
         // Clean up content objects
         _contentObjects.setNext(null);
